+++ examples/has_many_with_arguments.rs
@@ -114,7 +114,7 @@ pub struct Context {
 impl juniper::Context for Context {}
 
 #[derive(Clone, EagerLoading)]
-#[eager_loading(context = "Context", error = "diesel::result::Error")]
+#[eager_loading(context = Context, error = diesel::result::Error)]
 pub struct User {
     user: models::User,
 }
@@ -126,7 +126,7 @@ impl UserFields for User {
 }
 
 #[derive(Clone, EagerLoading)]
-#[eager_loading(context = "Context", error = "diesel::result::Error")]
+#[eager_loading(context = Context, error = diesel::result::Error)]
 pub struct Country {
     country: models::Country,