+++ juniper-eager-loading-code-gen/src/lib.rs
@@ -16,11 +16,13 @@ mod derive_eager_loading;
 mod impl_load_from_for_diesel;
 
 use impl_load_from_for_diesel::Backend;
+use proc_macro_error::*;
 
 #[proc_macro_derive(
     EagerLoading,
     attributes(eager_loading, has_one, option_has_one, has_many, has_many_through)
 )]
+#[proc_macro_error]
 pub fn derive_eager_loading(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     derive_eager_loading::gen_tokens(input)
 }